requirements.txt
# flask 
Werkzeug==1.0.1
MarkupSafe==1.1.1
Jinja2==2.11.2
itsdangerous==1.1.0
click==7.1.2
flask==1.1.2


# flask-SQLAlchemy
SQLAlchemy==1.3.20
flask-sqlalchemy==2.4.4

psycopg2==2.8.6


# migrations
Mako==1.1.3 
alembic==1.4.3 
flask-Migrate==2.5.3 
python-dateutil==2.8.1 
python-editor==1.0.4 
six==1.15.0


# authentication and authorization
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.2.0
rsa==4.6